[#170] refactor : 정보 공유하기/수정하기 리팩토링 완료
