+++ src/pages/item/create/index.tsx
@@ -1,13 +1,10 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useLayoutEffect, useState } from 'react'
 import BrandItemField from './components/BrandItemField/BrandItemField'
 import { useRecoilState, useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil'
 import DatePlaceField from './components/DatePlaceField/DatePlaceField'
 import PriceField from './components/PriceField/PriceField'
 import { useLocation, useNavigate, useParams } from 'react-router-dom'
-import SelectCeleb, {
-  selectedCelebState,
-  selectedGroupState,
-} from '../../../components/SelectCeleb/SelectCeleb'
+import SelectCeleb from '../../../components/SelectCeleb/SelectCeleb'
 import SelectCategory from './components/SelectCategory/SelectCategory'
 import Header from '../../../components/Header/Header'
 import {
@@ -31,14 +28,22 @@ import {
   parentCategoryState,
   subCategoryState,
 } from '../../../components/BottomSheetModal/ItemCategoryModal'
-import { linksState } from '../addLink/components/LinkInput/LinkInput'
 import ImageField from './components/ImageField/ImageField'
 import {
-  ICategory,
-  IHashTag,
-  IItemInfo,
+  createItemAddInfoState,
+  createItemBrandState,
+  createItemCategoryState,
   createItemCelebState,
-  itemInfoState,
+  createItemLinkState,
+  createItemNameState,
+  createItemNewBrandState,
+  createItemNewCelebState,
+  createItemPlaceState,
+  createItemPriceState,
+  createItemSourceState,
+  createItemWhenDateState,
+  currentTempIdState,
+  itemS3ImgListState,
 } from '../../../recoil/itemInfo'
 import useUploadStateObserver from '../../../hooks/useUploadStateObserver'
 import { localStorageKeys } from '../../../config/localStorageKeys'
@@ -47,123 +52,171 @@ import { modals } from '../../../components/Modals'
 import useItemQuery from '../../../apis/item/hooks/useItemQuery'
 import { HashTagResult, ImgResult, TempItemReq } from '../../../apis/item/itemService.type'
 import useItemDetailQuery from '../../../apis/item/hooks/useItemDetailQuery'
-import { Image, imgListState } from '../../../components/AddPhotos/AddPhotos'
+import { imgListState } from '../../../components/AddPhotos/AddPhotos'
 import useTempItemQuery from '../../../apis/item/hooks/useTempItemQuery'
+import { hashTagState } from '../addInfo/components/HashTags/HashTag'
+import useItemImgUpload from '../../../apis/s3/hooks/useItemImgUpload'
+import { checkListState } from '../temporary-storage'
 
-const ItemCreate = () => {
+const ItemCreateCopy = () => {
   useUploadStateObserver()
 
-  const { openModal } = useModals()
   const navigate = useNavigate()
-  const celeb = useRecoilValue(selectedCelebState)
-  const category = useRecoilValue(subCategoryState)
-  const [itemInfo, setItemInfo] = useRecoilState(itemInfoState)
-  const resetItemInfo = useResetRecoilState(itemInfoState)
-  const [hasTriedToUpload, setHasTriedToUpload] = useState(false)
+
+  const [hasTriedToUpload, setHasTriedToUpload] = useState<boolean>(false)
+
+  // Item 이미지 리스트
   const imgList = useRecoilValue(imgListState)
-  const setImgList = useSetRecoilState(imgListState)
+  const s3ImgList = useRecoilValue(itemS3ImgListState)
+  // Celeb ID
+  const celebInfoInItem = useRecoilValue(createItemCelebState)
+  // NewCeleb ID
+  const newCeleb = useRecoilValue(createItemNewCelebState)
+  // 착용 날짜
+  const whenDiscovery = useRecoilValue(createItemWhenDateState)
+  // 착용 장소
+  const whereDiscovery = useRecoilValue(createItemPlaceState)
+  // 아이템 카테고리
+  const category = useRecoilValue(createItemCategoryState)
+  // 브랜드 ID
+  const brand = useRecoilValue(createItemBrandState)
+  // NewBrand ID
+  const newBrand = useRecoilValue(createItemNewBrandState)
+  // 아이템 이름
+  const itemName = useRecoilValue(createItemNameState)
+  // 가격
+  const price = useRecoilValue(createItemPriceState)
+  // 추가정보
+  const additionalInfo = useRecoilValue(createItemAddInfoState)
+  const hashTags = useRecoilValue(hashTagState)
+  const source = useRecoilValue(createItemSourceState)
+  // 구매링크
+  const linkList = useRecoilValue(createItemLinkState)
+  // 임시저장 아이템
+  const [currentTempId, setCurrentTempId] = useRecoilState(currentTempIdState)
 
-  const resetCelebInfoInItem = useResetRecoilState(createItemCelebState)
-
-  const { getTempItem } = useTempItemQuery()
-  const { data } = getTempItem()
-
-  console.log('create render')
+  const { openModal } = useModals()
   useEffect(() => {
-    const newImgList: ImgResult[] = imgList.map((img: Image, idx) => ({
-      imgUrl: img.imgUrl ? img.imgUrl : '',
-      representFlag: idx === 0,
-      sortOrder: idx,
-    }))
-    if (imgList.length > 0) setItemInfo({ ...itemInfo, imgList: newImgList })
-  }, [imgList])
-
-  useEffect(() => {
-    const id = localStorage.getItem(localStorageKeys.TEMP_ITEM_ID)
-    if (id) {
+    console.log('useEffect')
+    const id = Number(localStorage.getItem(localStorageKeys.TEMP_ITEM_ID))
+    if (id && id !== currentTempId) {
       openModal(modals.AskRecentPostWritingModal)
     }
-    if (hasTriedToUpload === true) {
-      localStorage.removeItem(localStorageKeys.TEMP_ITEM_ID)
-    }
   }, [])
+
+  const resetS3ImgList = useResetRecoilState(itemS3ImgListState)
+  const resetImgListState = useResetRecoilState(imgListState)
+  const resetCelebInfoInItem = useResetRecoilState(createItemCelebState)
+  const resetNewCeleb = useResetRecoilState(createItemNewCelebState)
+  const resetWhenDiscovery = useResetRecoilState(createItemWhenDateState)
+  const resetwhereDiscovery = useResetRecoilState(createItemPlaceState)
+  const resetCategory = useResetRecoilState(createItemCategoryState)
+  const resetBrand = useResetRecoilState(createItemBrandState)
+  const resetNewBrand = useResetRecoilState(createItemNewBrandState)
+  const resetItemName = useResetRecoilState(createItemNameState)
+  const resetPrice = useResetRecoilState(createItemPriceState)
+  const resetAdditionalInfo = useResetRecoilState(createItemAddInfoState)
+  const resetHashTags = useResetRecoilState(hashTagState)
+  const resetSource = useResetRecoilState(createItemSourceState)
+  const resetLinkList = useResetRecoilState(createItemLinkState)
+
+  // const resetCategory = useResetRecoilState(subCategoryState)
+  // const resetParentCategory = useResetRecoilState(parentCategoryState)
+
+  const onBackClick = () => {
+    setCurrentTempId(null)
+    resetS3ImgList()
+    resetImgListState()
+    resetCelebInfoInItem()
+    resetNewCeleb()
+    resetWhenDiscovery()
+    resetwhereDiscovery()
+    resetCategory()
+    resetBrand()
+    resetNewBrand()
+    resetItemName()
+    resetPrice()
+    resetAdditionalInfo()
+    resetHashTags()
+    resetSource()
+    resetLinkList()
+    navigate('/home', { replace: true })
+  }
+
+  const { getTempCount, getTempItem } = useTempItemQuery()
+  const { data: tempCount } = getTempCount()
+  const { data: tempList } = getTempItem()
+
+  const {
+    postItemImg: { mutate: mutateByImgUpload },
+  } = useItemImgUpload()
+
+  const onSubmit = async () => {
+    const hashTagIdList: Array<number> | null =
+      hashTags && hashTags.length > 0 ? hashTags.map((item) => item.hashtagId) : null
+    setHasTriedToUpload(true)
+    // 아무 값이 없을 때
+    if (
+      !(
+        imgList ||
+        celebInfoInItem ||
+        whenDiscovery ||
+        whereDiscovery ||
+        category ||
+        brand ||
+        itemName ||
+        price ||
+        additionalInfo ||
+        hashTagIdList ||
+        linkList ||
+        source ||
+        newCeleb ||
+        newBrand
+      )
+    ) {
+      return
+    } else if (
+      imgList &&
+      (celebInfoInItem || newCeleb) &&
+      category &&
+      (brand?.brandId || newBrand?.brandId) &&
+      itemName &&
+      price
+    ) {
+      mutateByImgUpload(imgList)
+    }
+  }
 
   const {
     postItem: { mutate },
   } = useItemQuery()
 
-  const onSubmit = () => {
-    setHasTriedToUpload(true)
-    if (
-      itemInfo.imgList &&
-      (itemInfo.celeb || itemInfo.newCeleb) &&
-      itemInfo.itemCategory &&
-      (itemInfo.brand?.brandId || itemInfo.newBrand?.brandId) &&
-      itemInfo.itemName &&
-      itemInfo.price
-    ) {
-      let hashTags: Array<number> | null = []
-      if ((itemInfo.hashTagList?.length ?? 0) > 0) {
-        itemInfo?.hashTagList?.map((item) => {
-          if (item.hashtagId) hashTags?.push(item.hashtagId)
-        })
-      } else {
-        hashTags = null
-      }
-
-      if (
-        !(
-          itemInfo.imgList ||
-          itemInfo.celeb ||
-          itemInfo.whenDiscovery ||
-          itemInfo.whereDiscovery ||
-          itemInfo.itemCategory ||
-          itemInfo.brand ||
-          itemInfo.itemName ||
-          itemInfo.price ||
-          itemInfo.additionalInfo ||
-          itemInfo.hashTagList ||
-          itemInfo.linkList ||
-          itemInfo.infoSource ||
-          itemInfo.newCeleb ||
-          itemInfo.newBrand
-        )
-      ) {
-        return
-      }
+  // s3ImgList의 변화 감지를 위한 useEffect
+  useEffect(() => {
+    if (s3ImgList && s3ImgList.length > 0) {
+      // s3ImgList가 업데이트 되었을 때 실행할 로직
+      const hashTagIdList: Array<number> | null =
+        hashTags && hashTags.length > 0 ? hashTags.map((item) => item.hashtagId) : null
       const item: TempItemReq = {
-        id: itemInfo.id ? itemInfo.id : null,
-        imgList: itemInfo.imgList,
-        celebId: itemInfo.celeb?.celebId ?? null,
-        whenDiscovery: itemInfo.whenDiscovery && (itemInfo.whenDiscovery as Date).toISOString(),
-        whereDiscovery: itemInfo.whereDiscovery === '' ? null : itemInfo.whereDiscovery,
-        categoryId: itemInfo.itemCategory?.categoryId ?? null,
-        brandId: itemInfo.brand?.brandId ?? null,
-        itemName: itemInfo.itemName === '' ? null : itemInfo.itemName,
-        price: itemInfo.price ?? null,
-        additionalInfo: itemInfo.additionalInfo === '' ? null : itemInfo.additionalInfo,
-        hashTagIdList: hashTags,
-        linkList: itemInfo.linkList,
-        infoSource: itemInfo.infoSource === '' ? null : itemInfo.infoSource,
-        newCelebId: itemInfo.newCeleb?.celebId ?? null,
-        newBrandId: itemInfo.newBrand?.brandId ?? null,
+        id: null,
+        imgList: s3ImgList ?? null,
+        celebId: celebInfoInItem?.soloId ?? null,
+        whenDiscovery: whenDiscovery && (whenDiscovery as Date).toISOString(),
+        whereDiscovery: whereDiscovery === '' ? null : whereDiscovery,
+        categoryId: category?.categoryId ?? null,
+        brandId: brand?.brandId ?? null,
+        itemName: itemName === '' ? null : itemName,
+        price: price ?? null,
+        additionalInfo: additionalInfo ?? null,
+        hashTagIdList: hashTagIdList ?? null,
+        linkList: linkList ?? null,
+        infoSource: source ?? null,
+        newCelebId: newCeleb?.id ?? null,
+        newBrandId: newBrand?.brandId ?? null,
       }
       mutate(item)
     }
-  }
-
-  const resetCategory = useResetRecoilState(subCategoryState)
-  const resetParentCategory = useResetRecoilState(parentCategoryState)
-  const resetImgListState = useResetRecoilState(imgListState)
-
-  const onBackClick = () => {
-    resetItemInfo()
-    resetCelebInfoInItem()
-    resetCategory()
-    resetParentCategory()
-    resetImgListState()
-    navigate('/home', { replace: true })
-  }
+  }, [s3ImgList])
 
   return (
     <ItemCreatePageStyle>
@@ -176,16 +229,15 @@ const ItemCreate = () => {
       </HeaderWrapper>
       <ComponentContainer>
         <ComponentWrapper>
-          {/* <ImageField error={false}></ImageField> */}
           <ImageField hasTriedToUpload={hasTriedToUpload}></ImageField>
         </ComponentWrapper>
-        {/* <ComponentWrapper>
+        <ComponentWrapper>
           <LabelContainer>
-            {hasTriedToUpload && !celeb.id && <Error></Error>}
+            {hasTriedToUpload && !(celebInfoInItem?.soloId || newCeleb?.id) && <Error></Error>}
             <Label>누가 착용했나요?</Label>
           </LabelContainer>
           <SelectCeleb></SelectCeleb>
-          {hasTriedToUpload && !celeb.id && (
+          {hasTriedToUpload && !(celebInfoInItem?.soloId || newCeleb?.id) && (
             <ErrorText className='error'>필수 항목입니다</ErrorText>
           )}
         </ComponentWrapper>
@@ -198,38 +250,36 @@ const ItemCreate = () => {
         <ComponentWrapper>
           <LabelContainer>
             {hasTriedToUpload &&
-              (!category.id ||
-                (itemInfo.brand?.brandId && itemInfo.newBrand) ||
-                !itemInfo.itemName ||
-                !itemInfo.price) && <Error></Error>}
+              (!category?.categoryId || !(brand?.brandId || newBrand) || !itemName || !price) && (
+                <Error></Error>
+              )}
             <Label>어떤 아이템인가요?</Label>
           </LabelContainer>
           <SelectCategory />
-          {hasTriedToUpload && !category.id && (
+          {hasTriedToUpload && !category?.categoryId && (
             <ErrorText className='error'>필수 항목입니다</ErrorText>
           )}
           <ComponentWrapper className='padding'>
             <BrandItemField
-              brandValid={hasTriedToUpload ? !itemInfo.brand || !itemInfo.newBrand : true}
-              itemNameValid={hasTriedToUpload ? itemInfo.itemName !== '' : true}
+              brandValid={hasTriedToUpload ? !(!brand && !newBrand) : true}
+              itemNameValid={hasTriedToUpload ? itemName === '' && itemName === null : true}
             ></BrandItemField>
           </ComponentWrapper>
-          {(itemInfo.brand?.brandId || itemInfo.newBrand?.brandId) && (
+          {(brand?.brandId || newBrand?.brandId) && (
             <>
               <PriceField></PriceField>
-              {hasTriedToUpload && !itemInfo.price && (
+              {hasTriedToUpload && !price && (
                 <ErrorText className='error'>필수 항목입니다</ErrorText>
               )}
             </>
           )}
-        </ComponentWrapper> */}
+        </ComponentWrapper>
       </ComponentContainer>
 
-      {/* <BottomBar>
+      <BottomBar>
         <div className='left'>
           <div className='button' onClick={() => navigate('/item/create/addInfo')}>
-            {(itemInfo.additionalInfo && itemInfo.additionalInfo?.length > 0) ||
-            (itemInfo.hashTagList && itemInfo.hashTagList.length > 0) ? (
+            {(additionalInfo && additionalInfo?.length > 0) || (hashTags && hashTags.length > 0) ? (
               <InfoAddOn></InfoAddOn>
             ) : (
               <InfoAddOff></InfoAddOff>
@@ -238,24 +288,20 @@ const ItemCreate = () => {
             <span>추가 정보</span>
           </div>
           <div className='button' onClick={() => navigate('/item/create/addlink')}>
-            {itemInfo.linkList && itemInfo.linkList.length > 0 ? (
-              <LinkAddOn></LinkAddOn>
-            ) : (
-              <LinkAddOff></LinkAddOff>
-            )}
+            {linkList && linkList.length > 0 ? <LinkAddOn></LinkAddOn> : <LinkAddOff></LinkAddOff>}
             <span>구매 링크</span>
           </div>
         </div>
         <div className='right'>
-          {data?.pages[0].countNum && data?.pages[0].countNum > 0 ? (
+          {tempCount && tempCount?.tempItemCount > 0 ? (
             <StorageOn onClick={() => navigate('/item/create/temporary-storage')}></StorageOn>
           ) : (
             <StorageOff onClick={() => navigate('/item/create/temporary-storage')}></StorageOff>
           )}
         </div>
-      </BottomBar> */}
+      </BottomBar>
     </ItemCreatePageStyle>
   )
 }
 
-export default ItemCreate
+export default ItemCreateCopy
